Worked more on the internal json system
