+++ domains/attack_vector/encrypt.go
@@ -1,4 +1,4 @@
-package encrypt
+package attack_vector
 
 import (
 	"bufio"
