+++ domains/malware/parser.go
@@ -1,0 +1,74 @@
+package malware
+
+import (
+	"regexp"
+
+	"github.com/s9rA16Bf4/go-evil/utility/contains"
+	"github.com/s9rA16Bf4/go-evil/utility/error"
+	"github.com/s9rA16Bf4/go-evil/utility/json"
+)
+
+const (
+	EXTRACT_SUBDOMAIN = "(malware|#object|#self|#this)\\.(.+)\\.(.+)\\(.*\\);" // Captures subdomain and function
+	EXTRACT_FUNCTION  = "(malware|#object|#self|#this)\\.(.+)\\(.*\\);"        // This is for the cases when we don't have a subdomain
+)
+
+func Parse(base_64_serialize_json string) string {
+	data_structure := json.Receive(base_64_serialize_json)
+	data_structure.Append_to_call("Malware.Parse()")
+
+	for _, line := range data_structure.File_gut {
+		value := contains.Passed_value(line)
+		regex := regexp.MustCompile(EXTRACT_SUBDOMAIN)
+		result := regex.FindAllStringSubmatch(line, -1)
+
+		if len(result) > 0 { // There is a subdomain to extract
+			subdomain := result[0][2]
+			function := result[0][3]
+
+			switch subdomain {
+			case "set":
+				switch function {
+				case "extension":
+					data_structure = json.Receive(SetExtension(value, json.Send(data_structure)))
+				case "name":
+					data_structure = json.Receive(SetBinaryName(value, json.Send(data_structure)))
+				default:
+					error.Function_error(function, "malware.Parse()")
+				}
+			case "add":
+				switch function {
+				case "random_function":
+					data_structure = json.Receive(AddRandomFunction(json.Send(data_structure)))
+				case "random_variable":
+					data_structure = json.Receive(AddRandomVariable(json.Send(data_structure)))
+				case "static_file":
+					data_structure.Append_File_domain("runtime")
+					data_structure = json.Receive(AddStaticFile(value, json.Send(data_structure)))
+
+				default:
+					error.Function_error(function, "malware.Parse()")
+				}
+			case "disable":
+				switch function {
+				case "region":
+					data_structure = json.Receive(Disable_region(value, json.Send(data_structure)))
+				}
+			default:
+				error.Subdomain_error(subdomain, "malware.Parse()")
+			}
+		} else { // There might be a function which doesn't require a subdomain to work
+			regex = regexp.MustCompile(EXTRACT_FUNCTION)
+			result = regex.FindAllStringSubmatch(line, -1)
+			if len(result) > 0 {
+				function := result[0][2]
+				switch function {
+				default:
+					error.Function_error(function, "malware.Parse()")
+				}
+			}
+		}
+	}
+
+	return json.Send(data_structure)
+}
