+++ utility/parser/parser_regex.go
@@ -20,8 +20,11 @@ func Regex(base_64_serialize_json string) string {
 
 	for _, line := range data_structure.File_gut {
 		result := regex.FindAllStringSubmatch(line, -1)
-		data_structure.Append_File_domain(result[0][1])
 
+		if len(result) > 0 {
+			data_structure.Append_File_domain(result[0][1])
+
+		}
 	}
 
 	return json.Send(data_structure)
